# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the current user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1234567890abcdef12345678",
                    "username": "testuser",
                    "email": "test@example.com",
                    "createdAt": "2023-10-27T10:00:00.000Z",
                    "updatedAt": "2023-10-27T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateProfile",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              },
              "examples": {
                "user": {
                  "value": {
                    "username": "newusername",
                    "email": "newemail@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updates the current user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1234567890abcdef12345678",
                    "username": "newusername",
                    "email": "newemail@example.com",
                    "createdAt": "2023-10-27T10:00:00.000Z",
                    "updatedAt": "2023-10-27T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me/password": {
      "patch": {
        "operationId": "UsersController_changePassword",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              },
              "examples": {
                "user": {
                  "value": {
                    "currentPassword": "oldpassword",
                    "newPassword": "newpassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Changes the current user password.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1234567890abcdef12345678",
                    "username": "testuser",
                    "email": "test@example.com",
                    "createdAt": "2023-10-27T10:00:00.000Z",
                    "updatedAt": "2023-10-27T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              },
              "examples": {
                "user": {
                  "value": {
                    "username": "newadmin",
                    "email": "admin@example.com",
                    "password": "adminpassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a new user.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1234567890abcdef12345679",
                    "username": "newadmin",
                    "email": "admin@example.com",
                    "createdAt": "2023-10-27T10:00:00.000Z",
                    "updatedAt": "2023-10-27T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by field",
            "schema": {
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by name (partial match)",
            "schema": {
              "example": "John",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of all users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "1234567890abcdef12345678"
                      },
                      "username": {
                        "type": "string",
                        "example": "testuser"
                      },
                      "email": {
                        "type": "string",
                        "example": "test@example.com"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2023-10-27T10:00:00.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2023-10-27T10:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS SaaS API",
    "description": "API documentation for the LMS SaaS platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Jane Doe"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "Current password",
            "example": "oldpassword"
          },
          "newPassword": {
            "type": "string",
            "description": "New password (min 6 chars)",
            "example": "newpassword123"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Jane Doe"
          },
          "password": {
            "type": "string",
            "description": "User password (optional for system-generated users)",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "fullName"
        ]
      },
...
    }
  }
}